backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2.16
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-auth==0.9.3
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
